Game

+++ src/app/app.component.ts
@@ -1,15 +1,23 @@
-import { Component, HostListener, OnDestroy, OnInit } from "@angular/core";
 import { Clipboard } from "@angular/cdk/clipboard";
-import { Observable, of } from "rxjs";
-import { filter, take } from "rxjs/operators";
+import { Component, HostListener, OnDestroy, OnInit } from "@angular/core";
+import { Observable, Subject, of } from "rxjs";
+import {
+  filter,
+  switchMap,
+  take,
+  takeUntil,
+  withLatestFrom,
+} from "rxjs/operators";
 import { BehaviorSubject } from "rxjs/Rx";
 import { User } from "./auth/user.model";
+import { DialogService } from "./dialogs/dialog.service";
+import { ActionsTypes } from "./dialogs/rooms-dialog/rooms-dialog.component";
 import { LotteryService } from "./lottery.service";
 import { Element } from "./models/element";
 import { PeriodicTableService } from "./periodic-table.service";
-import { RoomsService } from "./rooms/rooms.service";
-import { UserService } from "./users/users.service";
+import { Room, RoomsService } from "./rooms/rooms.service";
 import { SnackBarService } from "./snackbar/snackbar.service";
+import { UserService } from "./users/users.service";
 
 @Component({
   selector: "app-root",
@@ -33,12 +41,20 @@ export class AppComponent implements OnInit, OnDestroy {
   rooms$ = this.roomsService.rooms$;
   myRoom$ = this.roomsService.myRoom$;
 
-  @HostListener("window:unload", ["$event"])
+  roomsDialog: any;
+
+  @HostListener("window:beforeunload", ["$event"])
   unloadHandler(event) {
+    this.userService.updateRoomAndRole(null, null);
+    console.log("DELETE VISITORs");
     this.userService.deleteVisitor();
     this.userData$.pipe(take(1)).subscribe((user) => {
-      if (user.role === "host") {
-        this.roomsService.removeRoom(user.room);
+      if (user && user.room) {
+        if (user.role === "host") {
+          this.roomsService.removeRoom(user.room);
+        } else {
+          this.roomsService.removeGuestFromRoom(user.room);
+        }
       }
     });
   }
@@ -51,57 +67,38 @@ export class AppComponent implements OnInit, OnDestroy {
     private roomsService: RoomsService,
     private lotteryService: LotteryService,
     private snackBarService: SnackBarService,
-    private clipboard: Clipboard
+    private clipboard: Clipboard,
+    private dialogService: DialogService
   ) {
     this.myRoom$.subscribe(({ searchingElement, startGame }) => {
       // TO DO: Unsubscribe
       this.searchingElement = searchingElement;
       this.level$.next("a");
-    });
-  }
 
-  storeToDB() {
-    this.userService.updateScore(this.state.score);
-  }
-
-  createRoom() {
-    this.userData$.pipe(take(1)).subscribe((user) => {
-      if (!user) {
-        this.snackBarService.openSnackBar("Login or play as guest");
-        return;
+      console.log(startGame);
+      if (startGame) {
+        this.roomsDialog.close();
       }
-      if (user.room) {
-        this.snackBarService.openSnackBar(
-          "You already have room, invite someone to play with you"
-        );
-        return;
+    });
+  }
+
+  hostPlayer$ = this.myRoom$.pipe(
+    switchMap<Room, Observable<User>>((myRoom) => {
+      return this.userService.getUser(myRoom.hostUserUid);
+    })
+  );
+  guestPlayer$ = this.myRoom$.pipe(
+    switchMap<Room, Observable<Partial<User>>>(({ guestUserUid }) => {
+      if (!guestUserUid) {
+        return of({
+          displayName: "trol",
+          score: -10,
+        });
       }
-      const room = this.roomsService.createRoom(user);
-      this.userService.updateRoomAndRole(room.key, "host");
-    });
-  }
 
-  joinRoom(roomKey: string) {
-    this.roomsService.joinRoom(roomKey);
-    this.userService.updateRoomAndRole(roomKey, "guest");
-  }
-
-  invitePlayer(roomKey: string) {
-    this.clipboard.copy("text");
-    this.snackBarService.openSnackBar("Send this link to your friend");
-  }
-
-  handleFinish(finish) {
-    this.userData$.pipe(take(1)).subscribe(({ room }) => {
-      this.roomsService.startGame(room, false, this.searchingElement);
-    });
-
-    this.state.start = finish;
-    if (this.state.score >= this.state.bestScore) {
-      this.state.bestScore = this.state.score;
-    }
-    this.state.score = 0;
-  }
+      return this.userService.getUser(guestUserUid);
+    })
+  );
 
   ngOnInit() {
     this.periodicTableService
@@ -112,6 +109,52 @@ export class AppComponent implements OnInit, OnDestroy {
 
     // TO DO: Unsubscribe
     this.userService.userApp$.subscribe();
+
+    this.playAsGuest();
+  }
+
+  private createRoom(user: User) {
+    if (!user) {
+      this.snackBarService.openSnackBar("Login or play as guest");
+      return;
+    }
+    if (user.room) {
+      this.snackBarService.openSnackBar(
+        "You already have room, invite someone to play with you"
+      );
+      return;
+    }
+    const room = this.roomsService.createRoom(user.uid, user.displayName);
+    this.userService.updateRoomAndRole(room.key, "host");
+  }
+
+  private joinRoom(roomKey: string, user: User) {
+    if (user.room) {
+      this.snackBarService.openSnackBar(
+        "You already have room, invite someone to play with you"
+      );
+      return;
+    }
+    this.roomsService.joinRoom(roomKey, user.uid);
+    this.userService.updateRoomAndRole(roomKey, "guest");
+  }
+
+  invitePlayer(roomKey: string) {
+    this.clipboard.copy(`https://chemgame.pl/room=${roomKey}`);
+    this.snackBarService.openSnackBar("Send this link to your friend");
+  }
+
+  handleFinish(finish) {
+    this.userData$.pipe(take(1)).subscribe(({ room }) => {
+      this.roomsService.startGame(room, false, this.searchingElement);
+      this.userService.updateBestScore(this.state.score);
+    });
+
+    this.state.start = finish;
+    if (this.state.score >= this.state.bestScore) {
+      this.state.bestScore = this.state.score;
+    }
+    this.state.score = 0;
   }
 
   handleSelected(symbol: string) {
@@ -124,6 +167,7 @@ export class AppComponent implements OnInit, OnDestroy {
       this.userData$.pipe(take(1)).subscribe(({ room }) => {
         this.searchingElement = this.randomElement();
         this.roomsService.searchingElement(room, this.searchingElement);
+        this.userService.updateScore(this.state.score);
       });
     }
   }
@@ -138,33 +182,60 @@ export class AppComponent implements OnInit, OnDestroy {
       }
       this.roomsService.startGame(user.room, true, this.searchingElement);
     });
+    console.log("START GAME");
     this.state.start = true;
-  }
-
-  get isPlaying(): Observable<boolean> {
-    return of(true);
-    // return combineLatest(this.myRoom$, this.userData$).pipe(
-    //   map(([{ player, startGame }, { role }]) => startGame && player !== role)
-    // );
-  }
-
-  get showActive(): Observable<boolean> {
-    return of(true);
-    // return combineLatest(this.myRoom$, this.userData$).pipe(
-    //   take(1),
-    //   map(([{ player }]) => player === "host")
-    // );
   }
 
   private randomElement(): string {
     return this.lotteryService.drawElement(this.state.score);
   }
 
-  playAsVisitor() {
-    this.userService.createVisitor();
+  removeRoom(roomKey: string) {
+    this.userService.updateRoomAndRole(null, null);
+    this.roomsService.removeRoom(roomKey);
   }
 
-  removeRoom(roomKey: string) {
-    return this.roomsService.removeRoom(roomKey);
+  playAsGuest() {
+    const dialogRef = this.dialogService.openWelcomDialog();
+    dialogRef
+      .afterClosed()
+      .pipe(filter(Boolean), take(1))
+      .subscribe((result) => {
+        this.userService.createVisitor(result as string);
+        this.showRooms();
+      });
+  }
+
+  showRooms() {
+    this.roomsDialog = this.dialogService.openRoomsDialog(this.rooms$);
+
+    this.roomsDialog.componentInstance.actions
+      .pipe(
+        withLatestFrom(this.userData$),
+        takeUntil(this.roomsDialog.afterClosed())
+      )
+      .subscribe(([{ key, type }, user]) => {
+        switch (type) {
+          case ActionsTypes.CREATE:
+            this.createRoom(user);
+            return;
+          case ActionsTypes.JOIN:
+            this.joinRoom(key, user);
+            return;
+          case ActionsTypes.DELETE:
+            if (user.room !== key) {
+              this.snackBarService.openSnackBar(
+                "You can delete only your own room"
+              );
+              return;
+            }
+            this.removeRoom(key);
+            return;
+          case ActionsTypes.START:
+            this.startGame();
+          default:
+            return;
+        }
+      });
   }
 }
